Collapse redundant redraw rectangles
Redraw is on a timer so with high refresh rates it batches,
which is cool, but if in that high refresh rate you send
the same triangle, blammo, on linux. This matters when you
have high resolution and high frequency mice. So put a uniqueness
check in the x11 frame invalid rectangle redraw code.