Handle OverDraws in Full Refresh case also
In the full refresh case (which surge uses a lot) we don't need
to also paint 300 sub children. So detect if the screen is indeed
doing that and force a single repaint. May be a more specific surge
optimization and really it should figure out a convex shell etc...
but this helps us