PNG Support from Raw Paths
PNG suport on Mac and Linux came from 'resources' which
were not the same as files; they either had to be memory
blocks (linux) or in the bundle (win). If you loaded with
a file path it blew up. This made our dynamic skin engine
from the file system hard to write.

So hack VSTGUI so if you ask for a file by path, it loads
the file by path, so you can display a file by path. Magic!